fix: split decision body coverage
