+++ packages/cli/test/decision-coverage-cli.test.ts
@@ -9,112 +9,6 @@ import { unwrapCommandReceipt } from "./helpers/receipt.ts";
 
 const cliEntry = path.resolve("packages/cli/src/index.ts");
 
-test("CLI decision propose and amend load authored markdown from --body-file", () => {
-  withTempRoot((rootDir) => {
-    const proposedBodyPath = path.join(rootDir, "proposed-body.md");
-    const amendedBodyPath = path.join(rootDir, "amended-body.md");
-    writeFileSync(proposedBodyPath, "## Background\n\nThe proposal needs prose.", "utf8");
-    writeFileSync(amendedBodyPath, "## Conclusion\n\nThe amended decision is clearer.", "utf8");
-
-    runJson(rootDir, [
-      "decision", "propose",
-      "--id", "dec_BODY_FILE",
-      "--title", "Body file input",
-      "--question", "Can a proposal read markdown from a file?",
-      "--chosen", "Read the file",
-      "--rejected", "Inline every paragraph",
-      "--why-not", "Long prose is difficult to quote safely",
-      "--body-file", proposedBodyPath
-    ]);
-
-    const proposedPath = path.join(rootDir, "harness/decisions/decision-dec_BODY_FILE/decision.md");
-    assert.match(decisionBody(readFileSync(proposedPath, "utf8")), /The proposal needs prose\./u);
-
-    runJson(rootDir, ["decision", "amend", "dec_BODY_FILE", "--body-file", amendedBodyPath]);
-    assert.match(decisionBody(readFileSync(proposedPath, "utf8")), /The amended decision is clearer\./u);
-  });
-});
-
-test("CLI decision body inputs reject --body and --body-file together", () => {
-  withTempRoot((rootDir) => {
-    const bodyPath = path.join(rootDir, "body.md");
-    writeFileSync(bodyPath, "File body", "utf8");
-
-    const result = runJson(rootDir, [
-      "decision", "propose",
-      "--title", "Conflicting body inputs",
-      "--question", "Should body inputs be mutually exclusive?",
-      "--chosen", "Reject the conflict",
-      "--rejected", "Guess which body wins",
-      "--why-not", "Silent precedence loses authored prose",
-      "--body", "Inline body",
-      "--body-file", bodyPath
-    ], false);
-
-    assert.equal(result.ok, false);
-    assert.equal(result.error?.code, "conflicting_decision_body_input");
-    assert.match(result.error?.hint ?? "", /only one of --body or --body-file/u);
-  });
-});
-
-test("CLI decision body input reports a missing --body-file clearly", () => {
-  withTempRoot((rootDir) => {
-    const missingPath = path.join(rootDir, "missing-body.md");
-    const result = runJson(rootDir, ["decision", "amend", "dec_MISSING_BODY_FILE", "--body-file", missingPath], false);
-
-    assert.equal(result.ok, false);
-    assert.equal(result.error?.code, "decision_body_file_read_failed");
-    assert.match(result.error?.hint ?? "", /Could not read decision body file/u);
-    assert.match(result.error?.hint ?? "", /missing-body\.md/u);
-  });
-});
-
-test("CLI decision propose and accept warn when markdown prose is empty, without blocking transitions", () => {
-  withTempRoot((rootDir) => {
-    const proposal = runJson(rootDir, [
-      "decision", "propose",
-      "--id", "dec_EMPTY_BODY",
-      "--title", "Empty prose warning",
-      "--question", "Should an empty decision body be visible?",
-      "--chosen", "Warn without blocking",
-      "--rejected", "Reject the write",
-      "--why-not", "A hard gate encourages placeholder prose",
-      "--evidence-relation", "C1:relates:task/task_01BODY:Evidence supports the claim."
-    ]);
-
-    assert.equal(proposal.ok, true);
-    assert.equal(proposal.warnings?.some((warning: { code?: string }) => warning.code === "decision_body_empty"), true);
-
-    const accepted = runJson(rootDir, ["decision", "accept", "dec_EMPTY_BODY", "--arbiter", "human:ZeyuLi"]);
-    assert.equal(accepted.ok, true);
-    assert.equal(accepted.decisionState, "active");
-    assert.equal(accepted.warnings?.some((warning: { code?: string }) => warning.code === "decision_body_empty"), true);
-  });
-});
-
-test("CLI decision proposal and acceptance omit the empty-body warning when prose exists", () => {
-  withTempRoot((rootDir) => {
-    const proposal = runJson(rootDir, [
-      "decision", "propose",
-      "--id", "dec_NONEMPTY_BODY",
-      "--title", "Written prose",
-      "--question", "Does written prose avoid the warning?",
-      "--chosen", "Do not warn",
-      "--rejected", "Warn anyway",
-      "--why-not", "The narrative is already present",
-      "--evidence-relation", "C1:relates:task/task_01BODY:Evidence supports the claim.",
-      "--body", "## Background\n\nThe reader can see why this choice exists."
-    ]);
-
-    assert.equal(proposal.ok, true);
-    assert.equal(proposal.warnings?.some((warning: { code?: string }) => warning.code === "decision_body_empty") ?? false, false);
-
-    const accepted = runJson(rootDir, ["decision", "accept", "dec_NONEMPTY_BODY", "--arbiter", "human:ZeyuLi"]);
-    assert.equal(accepted.ok, true);
-    assert.equal(accepted.warnings?.some((warning: { code?: string }) => warning.code === "decision_body_empty") ?? false, false);
-  });
-});
-
 test("CLI decision propose writes multiple claims with derived ids from JSON input", () => {
   withTempRoot((rootDir) => {
     const payload = JSON.stringify({
